[rename/WEDPRO-50] rename: 파일 구조 변경
