+++ src/App.jsx
@@ -6,7 +6,7 @@ import { Reset } from "styled-reset";
 // 페이지 불러오기
 //import Layout from "./pages/Layout";
 import Home from "./pages/Home";
-import Admin from "./pages/Admin";
+import Admin from "./pages/admin/Admin.jsx";
 // 절대경로 설정 예시 // @사용
 import NotFound from "@/pages/NotFound";
 // import NotFound from './pages/NotFound'
@@ -15,11 +15,11 @@ import NotFound from "@/pages/NotFound";
 import Main from "./pages/main/MainPage";
 
 // 관리자 페이지
-import AdminAlarmLists from "./pages/AdminAlarmLists";
-import AdminAttendanceLists from "./pages/AdminAttendanceLists";
-import AdminDonationLists from "./pages/AdminDonationLists";
-import AdminMemo from "./pages/AdminMemo";
-import AdminMain from "./pages/AdminMain";
+import AdminAlarmLists from "./pages/admin/AdminAlarmLists.jsx";
+import AdminAttendanceLists from "./pages/admin/AdminAttendanceLists.jsx";
+import AdminDonationLists from "./pages/admin/AdminDonationLists.jsx";
+import AdminMemo from "./pages/admin/AdminMemo.jsx";
+import AdminMain from "./pages/admin/AdminMain.jsx";
 
 // 상품 갤러리
 import GalleryWedding from "./pages/gallery/GalleryWedding";
