+++ src/pages/admin/AdminAlarmLists.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import AdminAlarmListsContainer from "../../containers/admin/AdminAlarmListsContainer.jsx";
+
+// 알림 목록 확인 페이지
+function AdminAlarmLists() {
+  return (
+    <AdminAlarmListsContainer></AdminAlarmListsContainer>
+  )
+}
+
+export default AdminAlarmLists
