+++ src/pages/admin/AdminDonationLists.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import AdminDonationListsContainer from "../../containers/admin/AdminDonationListsContainer.jsx";
+
+
+
+function AdminDonationLists() {
+  return (
+   <AdminDonationListsContainer></AdminDonationListsContainer>
+  )
+}
+
+export default AdminDonationLists
